Added auth

+++ app/Http/Controllers/IdeaController.php
@@ -9,12 +9,15 @@ class IdeaController extends Controller
 {
     public function update(idea $idea)
     {
-        request()->validate([
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240',
         ]);
 
-        $idea->content = request()->get('content', "");
-        $idea->save();
+        $idea->update($validated);
 
         // return view('ideas.show',compact('idea', 'editing'));
         return redirect()->route('ideas.show', $idea->id)->with('success','Idea edited successfully!');
@@ -22,6 +25,10 @@ class IdeaController extends Controller
 
     public function edit(idea $idea)
     {
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $editing = true;
 
         return view('ideas.show',compact('idea', 'editing'));
@@ -29,6 +36,8 @@ class IdeaController extends Controller
 
     public function show(idea $idea)
     {
+        
+
         return view('ideas.show',compact('idea'));
         // return view('ideas.show',[
         //     'idea'=>$idea,
@@ -38,22 +47,24 @@ class IdeaController extends Controller
     public function store()
     {
 
-        request()->validate([
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240',
         ]);
 
-        idea::create(
-            [
-                'content' => request()->get('content',''),
-                // 'likes' => 0,
-            ]
-        );
+        $validated['user_id'] = auth()->id();
+
+        idea::create($validated);
+        // idea::create(request()->all());
 
         return redirect()->route('dashboard')->with('success','Idea created successfully!');
     }
 
     public function destroy(idea $idea)
     {
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
 
         $idea->delete();
         // idea::where('id',$id)->firstOrFail()->delete();
